requirements.txt
Flask==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.1.3
Jinja2==3.1.6
click==8.1.7
itsdangerous==2.2.0
MarkupSafe==3.0.2
python-dotenv==1.0.0
requests==2.31.0
boto3==1.35.84
botocore==1.35.84
Pillow==10.4.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
awsebcli==3.21.0
PyMySQL==1.1.0
mysql-connector-python==8.0.33
cryptography==42.0.5
pydub==0.25.1
ffmpeg-python==0.2.0
